+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/RecordProcessorStep.java
@@ -20,7 +20,7 @@
 package org.neo4j.unsafe.impl.batchimport;
 
 import org.neo4j.kernel.impl.store.record.AbstractBaseRecord;
-import org.neo4j.unsafe.impl.batchimport.staging.ExecutorServiceStep;
+import org.neo4j.unsafe.impl.batchimport.staging.ProcessorStep;
 import org.neo4j.unsafe.impl.batchimport.staging.StageControl;
 import org.neo4j.unsafe.impl.batchimport.staging.Step;
 import org.neo4j.unsafe.impl.batchimport.stats.StatsProvider;
@@ -28,7 +28,7 @@ import org.neo4j.unsafe.impl.batchimport.stats.StatsProvider;
 /**
  * {@link RecordProcessor} in {@link Step Step-form}.
  */
-public class RecordProcessorStep<T extends AbstractBaseRecord> extends ExecutorServiceStep<T[]>
+public class RecordProcessorStep<T extends AbstractBaseRecord> extends ProcessorStep<T[]>
 {
     private final RecordProcessor<T> processor;
     private final boolean endOfLine;
