+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/RelationshipPreparationStep.java
@@ -22,7 +22,7 @@ package org.neo4j.unsafe.impl.batchimport;
 import org.neo4j.kernel.impl.store.record.RelationshipRecord;
 import org.neo4j.unsafe.impl.batchimport.cache.idmapping.IdMapper;
 import org.neo4j.unsafe.impl.batchimport.input.InputRelationship;
-import org.neo4j.unsafe.impl.batchimport.staging.ExecutorServiceStep;
+import org.neo4j.unsafe.impl.batchimport.staging.ProcessorStep;
 import org.neo4j.unsafe.impl.batchimport.staging.StageControl;
 
 /**
@@ -30,7 +30,7 @@ import org.neo4j.unsafe.impl.batchimport.staging.StageControl;
  * This step is also parallelizable so if it becomes a bottleneck then more processors will automatically
  * be assigned to it.
  */
-public class RelationshipPreparationStep extends ExecutorServiceStep<Batch<InputRelationship,RelationshipRecord>>
+public class RelationshipPreparationStep extends ProcessorStep<Batch<InputRelationship,RelationshipRecord>>
 {
     private final IdMapper idMapper;
