+++ community/kernel/src/main/java/org/neo4j/unsafe/impl/batchimport/UpdateRecordsStep.java
@@ -23,7 +23,7 @@ import java.util.Collection;
 
 import org.neo4j.kernel.impl.store.RecordStore;
 import org.neo4j.kernel.impl.store.record.AbstractBaseRecord;
-import org.neo4j.unsafe.impl.batchimport.staging.ExecutorServiceStep;
+import org.neo4j.unsafe.impl.batchimport.staging.ProcessorStep;
 import org.neo4j.unsafe.impl.batchimport.staging.StageControl;
 import org.neo4j.unsafe.impl.batchimport.stats.Key;
 import org.neo4j.unsafe.impl.batchimport.stats.Keys;
@@ -34,7 +34,7 @@ import org.neo4j.unsafe.impl.batchimport.stats.StatsProvider;
  * Updates a batch of records to a store.
  */
 public class UpdateRecordsStep<RECORD extends AbstractBaseRecord>
-        extends ExecutorServiceStep<RECORD[]>
+        extends ProcessorStep<RECORD[]>
         implements StatsProvider
 {
     private final RecordStore<RECORD> store;
